refactor: move controller and service files to module root
