+++ src/modules/event/event.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { EventController } from './controller/event.controller';
-import { EventService } from './service/event.service';
+import { EventController } from './event.controller';
+import { EventService } from './event.service';
 
 @Module({
   controllers: [EventController],
